requirements.txt
python-dotenv
gradio
pypdf
openai